requirements.txt
langgraph
langchain-ollama
streamlit
python-dotenv
mcp
httpx
pydantic
pydantic-settings
